+++ dashboard/package.json
@@ -9,7 +9,8 @@
   },
   "dependencies": {
     "react": "^18.0.0",
-    "react-dom": "^18.0.0"
+    "react-dom": "^18.0.0",
+    "react-router-dom": "^6.13.0"
   },
   "devDependencies": {
     "@types/node": "^18.13.0",
